[MAT-288] Split out the parts of expression that are to be generated.
