+++ include/expression.hh
@@ -14,6 +14,7 @@
 #include <vector>
 #include <memory>
 
+#include "expressionbase.hh"
 #include "numeric.hh"
 #include "visitor.hh"
 #include "exceptions.hh"
@@ -27,49 +28,6 @@ using namespace std;
  */
 namespace librdag
 {
-
-/**
- * Container for expression arguments
- */
-
-typedef OwningPtrVector<const OGNumeric*> ArgContainer;
-typedef OwningPtrVector<const OGNumeric*>  RegContainer;
-
-/**
- *  Expr type
- */
-class OGExpr: public OGNumeric
-{
-  public:
-    virtual ~OGExpr();
-    const ArgContainer* getArgs() const;
-    size_t getNArgs() const;
-    virtual const OGExpr* asOGExpr() const override;    
-    virtual void debug_print() const override;
-    virtual void accept(Visitor &v) const override;
-    virtual const RegContainer * getRegs() const;
-  protected:
-    OGExpr(ArgContainer* args);
-  private:
-    ArgContainer * _args;
-    RegContainer * _regs;
-};
-
-/**
- * Things that extend OGExpr
- */
-
-class OGUnaryExpr: public OGExpr
-{
-  protected:
-    OGUnaryExpr(ArgContainer* args);
-};
-
-class OGBinaryExpr : public OGExpr
-{
-  protected:
-    OGBinaryExpr(ArgContainer* args);
-};
 
 class COPY: public OGUnaryExpr
 {
